Fix iOS build and handle stubbed epoch-millis timestamps on Android
iOS: the standalone RelativeTime.swift was not resolving in the app
target ("cannot find 'RelativeTime' in scope" at PostDetailView.swift),
since it was the only app file missing from the test-target membership
exception lists of the synchronized project. Move the RelativeTime enum
into PostDetailView.swift (the only consumer); the unit test still
reaches it via @testable import.

Android: parseBackendDate only handled ISO-8601, but the stubbed API
used in debug/UI-test builds emits creation_time as epoch-millis strings
(System.currentTimeMillis().toString()). Those failed to parse and
collapsed every comment time to "< 1 min". Parse a bare integer string
as epoch-millis (unambiguous, never valid ISO-8601) and short-circuit
blanks. Add tests for both.

Addresses Copilot review feedback on #244.
