+++ android/PositiveOnlySocial/app/src/main/java/com/example/positiveonlysocial/util/BackendDate.kt
@@ -5,15 +5,29 @@ import java.time.format.DateTimeParseException
 import java.util.Date
 
 /**
- * Parses an ISO-8601 timestamp produced by the Django backend — typically with
- * fractional seconds and a `+00:00` offset, e.g.
- * "2024-01-15T10:30:45.123456+00:00" — into a [Date].
+ * Parses a backend `creation_time` string into a [Date].
  *
- * Returns null when the string is blank or can't be parsed, so callers can
- * choose their own fallback. This mirrors the iOS `parseDate` helper.
+ * Handles both formats the app sees:
+ * - The real Django backend emits ISO-8601, typically with fractional seconds
+ *   and a `+00:00` offset, e.g. "2024-01-15T10:30:45.123456+00:00".
+ * - The stubbed API used in debug/UI-test builds emits epoch-millis strings,
+ *   e.g. "1718800000000" (see `StatefulStubbedAPI`).
+ *
+ * Returns null when the string is blank or can't be parsed in either format,
+ * so callers can choose their own fallback. This mirrors the iOS `parseDate`
+ * helper.
  */
-fun parseBackendDate(value: String): Date? = try {
-    Date.from(OffsetDateTime.parse(value).toInstant())
-} catch (_: DateTimeParseException) {
-    null
+fun parseBackendDate(value: String): Date? {
+    val trimmed = value.trim()
+    if (trimmed.isEmpty()) return null
+
+    // Epoch-millis (the stubbed API). A bare integer string is never valid
+    // ISO-8601, so this is unambiguous.
+    trimmed.toLongOrNull()?.let { return Date(it) }
+
+    return try {
+        Date.from(OffsetDateTime.parse(trimmed).toInstant())
+    } catch (_: DateTimeParseException) {
+        null
+    }
 }
